Optimize friends page handle default avatar for friends page properly
